+++ tests/modules/MintersIntegration.t.sol
@@ -19,18 +19,18 @@ contract MintersIntegration is TestConfig {
 
     // Free drop constant properties
     uint96 PRICE_FREE_DROP = 0;
-    uint24 MINTER_MAX_MINTABLE_FREE_DROP = 100;
-    uint24 MAX_MINTABLE_PER_ACCOUNT = 3;
+    uint32 MINTER_MAX_MINTABLE_FREE_DROP = 100;
+    uint32 MAX_MINTABLE_PER_ACCOUNT = 3;
 
     // Presale constant properties
     uint96 PRICE_PRESALE = 50000000 gwei; // Price is 0.05 ETH
-    uint24 MINTER_MAX_MINTABLE_PRESALE = 45;
-    uint24 MAX_MINTABLE_PER_ACCOUNT_PRESALE = 25; // There is a 25 tokens per account limit set on the presale.
+    uint32 MINTER_MAX_MINTABLE_PRESALE = 45;
+    uint32 MAX_MINTABLE_PER_ACCOUNT_PRESALE = 25; // There is a 25 tokens per account limit set on the presale.
 
     // Public sale constant properties
     uint96 PRICE_PUBLIC_SALE = 100000000000000000; // Price is 0.1 ETH
-    uint24 MINTER_MAX_MINTABLE_PUBLIC_SALE = 700;
-    uint24 MAX_MINTABLE_PER_ACCOUNT_PUBLIC_SALE = 50;
+    uint32 MINTER_MAX_MINTABLE_PUBLIC_SALE = 700;
+    uint32 MAX_MINTABLE_PER_ACCOUNT_PUBLIC_SALE = 50;
 
     address[] public userAccounts = [
         getFundedAccount(1), // User 1 - participate in free drop
@@ -68,7 +68,7 @@ contract MintersIntegration is TestConfig {
       - Public Sale: executed as 1 day long public sale
     */
     function test_Glasshouse() public {
-        uint24 EDITION_MAX_MINTABLE = 1000;
+        uint32 EDITION_MAX_MINTABLE = 1000;
         // Setup Glass house sound edition
         edition = SoundEditionV1(
             createSound(
